+++ packages/create-sitecore-jss/src/templates/nextjs-sxa/src/Layout.tsx
@@ -5,11 +5,11 @@ import React from 'react';
 import Head from 'next/head';
 import {
   Placeholder,
-  getPublicUrl,
   LayoutServiceData,
   Field,
   HTMLLink,
 } from '@sitecore-jss/sitecore-jss-nextjs';
+import { getPublicUrl } from '@sitecore-jss/sitecore-jss-nextjs/utils';
 import Scripts from 'src/Scripts';
 
 // Prefix public assets with a public URL to enable compatibility with Sitecore Experience Editor.