+++ packages/sitecore-jss-nextjs/src/middleware/index.ts
@@ -1,5 +1,3 @@
 export { RedirectsMiddleware, RedirectsMiddlewareConfig } from './redirects-middleware';
 export { PersonalizeMiddleware, PersonalizeMiddlewareConfig } from './personalize-middleware';
 export { MultisiteMiddleware, MultisiteMiddlewareConfig } from './multisite-middleware';
-export { SiteResolver, SiteInfo } from '@sitecore-jss/sitecore-jss/site';
-export { tryParseEnvValue } from '@sitecore-jss/sitecore-jss/utils';