+++ packages/sitecore-jss-nextjs/src/utils/index.ts
@@ -1,0 +1,7 @@
+export { getPublicUrl, handleEditorFastRefresh } from './utils';
+export {
+  tryParseEnvValue,
+  isEditorActive,
+  resetEditorChromes,
+  resolveUrl,
+} from '@sitecore-jss/sitecore-jss/utils';